Refactor and expose more features of date/time package
