--- packages/datetime/src/dbquery.jsonlogic.date-fns.test.ts
@@ -1,4 +1,0 @@
-import { runJsonLogicTests } from './dbqueryJsonLogicTestUtils';
-import { rqbDateTimeLibraryAPI } from './rqbDateTimeLibraryAPI.date-fns';
-
-runJsonLogicTests(rqbDateTimeLibraryAPI);
