--- packages/datetime/src/dbquery.jsonlogic.dayjs.test.ts
@@ -1,4 +1,0 @@
-import { runJsonLogicTests } from './dbqueryJsonLogicTestUtils';
-import { rqbDateTimeLibraryAPI } from './rqbDateTimeLibraryAPI.dayjs';
-
-runJsonLogicTests(rqbDateTimeLibraryAPI);
