--- packages/datetime/src/dbquery.jsonlogic.luxon.test.ts
@@ -1,4 +1,0 @@
-import { runJsonLogicTests } from './dbqueryJsonLogicTestUtils';
-import { rqbDateTimeLibraryAPI } from './rqbDateTimeLibraryAPI.luxon';
-
-runJsonLogicTests(rqbDateTimeLibraryAPI);
